// src/types/enum/restaurant-category.enum.ts
export enum RestaurantCategory {
  // 아시아 음식
  KOREAN = 'KOREAN', // 한식
  JAPANESE = 'JAPANESE', // 일식
  CHINESE = 'CHINESE', // 중식
  ASIAN = 'ASIAN', // 아시안

  // 서양 음식
  WESTERN = 'WESTERN', // 양식
  MEXICAN = 'MEXICAN', // 멕시칸

  // 기타
  CAFE = 'CAFE', // 카페
  DESSERT = 'DESSERT', // 디저트
  BAR = 'BAR', // 술집 or 펍
  BUFFET = 'BUFFET', // 뷔페
  FUSION = 'FUSION', // 퓨전
  VEGAN = 'VEGAN', // 비건
  FAST_FOOD = 'FAST_FOOD', // 패스트푸드
}